+++ src/routes/notifications.js
@@ -2,7 +2,7 @@ const express = require("express");
 const router = express.Router();
 const { query, body } = require("express-validator");
 
-const notificationHandlers = require("../service/notifications");
+const notificationHandlers = require("../services/notifications");
 const validator = require("../middlewares/validator");
 
 // 라우터 접근 시 로그인 필요
